Add baseurl to table for User & Project Pages
I had a bit of a hard time finding out what I should write as the baseurl in the _config.yml because the article in chapter 4 is only about Project Pages and the chapter about User Pages also does not exactly describe it. Through trial and error I found out. Thought it might be good to have it in the table for others to faster find the solution.